djangoproject/requirements.txt
asgiref==3.6.0
Django==4.1.4
gunicorn==20.1.0
pkg_resources==0.0.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
sqlparse==0.4.3
